feat(shader): separate shader defs per render bucket
When a RDR2 shader has multiple render buckets defined in the XML, now
it creates one ShaderDef per render bucket, where the preset name (aka
filename) is the base name + some suffix depending on the render bucket.

This behaviour is closer to what we were doing with the .sps files in
GTA5. Eventually, we should switch to using the .sps files found in the
RDR2 game files too (some preset names currently don't match the .sps
files, which is not really a problem as these are used only for UI).

On import, it now uses both the base name and render bucket to find the
correct ShaderDef, instead of just the base name.

Also, various fixes where the game was not specified when accessing
shaders and was always using GTA5.